base: Pregenerate stdlib_disallow.inc
The genrule is flaky with Bazel, and it's unclear why. Instead, just add
a normal script and get rid of the genrule.

This unfortunately requires us to manually run the
stdlib_disallow_update.sh script, but we don't do this too often. It's
easily spotted if we forget to run the script on additions to the
allow-list, because the build breaks. If we forget to run the script on
removals, let's hope the code reviewer will notice (famous last words).

PiperOrigin-RevId: 640154125