build: Remove self features from libraries and tests
When setting up a Bazel Toolchain Configuration that provides features to
directly compile libraries with compiler metadata and link the GWPSan runtime,
we want to support compiling GWPSan itself with minimal changes.

A reasonable name for these features are "gwpsan_metadata" (for compiler
metadata) and "gwpsan_runtime" (for linking the runtime).

Remove both features from all internal libraries and binaries.

PiperOrigin-RevId: 660018257