Warn more carefully if time goes backwards
Generally monotonic time going backwards should not happen
provided the kernel is not broken. However, it can happen in some process
migration cases if the migration procedure does not use time namespaces
to restore the monotonic clock.
On one hand, we want to warn about this, but on the other hand, we don't
want to warn too often, and don't want to warn in migration cases.
So we warn once if this happens twice in a row.

PiperOrigin-RevId: 802472302